refactor(backdrop)
Now on both JP and CJ interfaces and as its own module.
Better text for CJ backdrop.

Additional work:
Comments in templates.
Fix nesting level of result panel in JP interface.
Fix enableScoring not called when reloading JP while a match is in
progress.
Allow scoringStateChanged events to be sent to unauthorised CJs.