review identification mechanism
Client passes its identity during the connection process (and soon its
session ID), which removes the need for an identity confirmation
process.

If the client passes a session ID as well, the server finds the
corresponding user and restore its session. If no session ID is passed,
the identification process (password/name) is performed, then, if the
identification is successful, the server initialises the new user and
sends its ID to the client. The client then needs to store this ID in a
cookie and send it the next time it tries to connect (after a page
refresh for instance).

TODO: save ID in cookie on the client, and send it when connecting.