ensure totals are computed when displaying results
If CJ joins after a period's rounds have ended, but before the next
period has started, a scoring sheet is created for him/her for that
period, for which the totals never get computed. This has no impact and
doesn't need to be fixed, apart from the way it is displayed in the
results table. With this change, if the totals are not computed, display
'-' in the table instead of leaving the cells empty.